[SDK-201] Additional runtime checks for logged in status

+++ src/main/kotlin/us/frollo/frollosdk/bills/Bills.kt
@@ -21,10 +21,14 @@ import androidx.lifecycle.Transformations
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.uiThread
 import us.frollo.frollosdk.aggregation.Aggregation
+import us.frollo.frollosdk.authentication.Authentication
 import us.frollo.frollosdk.base.Resource
 import us.frollo.frollosdk.base.Result
 import us.frollo.frollosdk.core.OnFrolloSDKCompletionListener
 import us.frollo.frollosdk.database.SDKDatabase
+import us.frollo.frollosdk.error.DataError
+import us.frollo.frollosdk.error.DataErrorSubType
+import us.frollo.frollosdk.error.DataErrorType
 import us.frollo.frollosdk.extensions.enqueue
 import us.frollo.frollosdk.extensions.fetchBillPayments
 import us.frollo.frollosdk.network.NetworkService
@@ -47,7 +51,7 @@ import us.frollo.frollosdk.network.api.BillsAPI
 import java.math.BigDecimal
 
 /** Manages bills and bill payments */
-class Bills(network: NetworkService, private val db: SDKDatabase, private val aggregation: Aggregation) {
+class Bills(network: NetworkService, private val db: SDKDatabase, private val aggregation: Aggregation, private val authentication: Authentication) {
 
     companion object {
         private const val TAG = "Bills"
@@ -160,6 +164,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
     }
 
     private fun createBill(request: BillCreateRequest, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#createBill", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         billsAPI.createBill(request).enqueue { resource ->
             when (resource.status) {
                 Resource.Status.ERROR -> {
@@ -180,6 +191,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun deleteBill(billId: Long, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#deleteBill", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         billsAPI.deleteBill(billId).enqueue { resource ->
             when (resource.status) {
                 Resource.Status.ERROR -> {
@@ -201,6 +219,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun refreshBills(completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#refreshBills", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         billsAPI.fetchBills().enqueue { resource ->
             when (resource.status) {
                 Resource.Status.ERROR -> {
@@ -221,6 +246,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun refreshBill(billId: Long, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#refreshBill", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         billsAPI.fetchBill(billId).enqueue { resource ->
             when (resource.status) {
                 Resource.Status.ERROR -> {
@@ -241,6 +273,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun updateBill(bill: Bill, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#updateBill", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         val request = BillUpdateRequest(
                 name = bill.name,
                 billType = bill.billType,
@@ -362,6 +401,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun deleteBillPayment(billPaymentId: Long, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#deleteBillPayment", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         billsAPI.deleteBillPayment(billPaymentId).enqueue { resource ->
             when (resource.status) {
                 Resource.Status.ERROR -> {
@@ -383,6 +429,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun refreshBillPayments(fromDate: String, toDate: String, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#refreshBillPayments", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         billsAPI.fetchBillPayments(fromDate = fromDate, toDate = toDate).enqueue { resource ->
             when (resource.status) {
                 Resource.Status.ERROR -> {
@@ -405,6 +458,13 @@ class Bills(network: NetworkService, private val db: SDKDatabase, private val ag
      * @param completion Optional completion handler with optional error if the request fails
      */
     fun updateBillPayment(billPaymentId: Long, date: String? = null, paid: Boolean? = null, completion: OnFrolloSDKCompletionListener<Result>? = null) {
+        if (!authentication.loggedIn) {
+            val error = DataError(type = DataErrorType.AUTHENTICATION, subType = DataErrorSubType.LOGGED_OUT)
+            Log.e("$TAG#updateBillPayment", error.localizedDescription)
+            completion?.invoke(Result.error(error))
+            return
+        }
+
         val status = paid?.let {
             if (paid) BillPaymentRequestStatus.PAID else BillPaymentRequestStatus.UNPAID
         }
